requirements.txt
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.2
plotly==5.15.0
python-dotenv==1.0.0
streamlit==1.25.0
jupyter==1.0.0